Add certificate chain verification for card authenticity

+++ libraries/Adafruit_PN532/Adafruit_PN532.h
@@ -167,7 +167,7 @@ public:
   bool startPassiveTargetIDDetection(uint8_t cardbaudrate);
   bool readDetectedPassiveTargetID(uint8_t *uid, uint8_t *uidLength);
   bool inDataExchange(uint8_t *send, uint8_t sendLength, uint8_t *response,
-                      uint8_t *responseLength);
+                      uint16_t *responseLength);
   bool inListPassiveTarget();
   uint8_t AsTarget();
   uint8_t getDataTarget(uint8_t *cmd, uint8_t *cmdlen);
